pyproject.toml
[tool.poetry]
name = "smartwater"
version = "0.1.0"
description = ""
authors = ["my <jakha921@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.5"
djangorestframework-simplejwt = "^5.2.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
